theme switching

+++ OpenPlex 0.2/AppDelegate.applescript
@@ -430,7 +430,14 @@ script AppDelegate
     on buttonhandlerrestoreall_(sender)
         tell myProgressBar to startAnimation:me -- another way
         set animated to true
-        do shell script "rm /Applications/PlexConnect/assets/fanartcache/*"
+        tell application "Finder"
+            if (exists folder "Applications:PlexConnect:assets:fanartcache" of the startup disk) then
+                do shell script "rm /Applications/PlexConnect/assets/fanartcache/*"
+                else if not (exists folder "Applications:PlexConnect:assets:fanartcache" of the startup disk) then
+                do shell script "echo already present"
+            end if
+        end tell
+        
         do shell script "cp /Applications/plexconnect_BACKUP/ATVSettings.cfg /Applications/PlexConnect"
         do shell script "cp /Applications/plexconnect_BACKUP/trailers.cer /Applications/PlexConnect/assets/certificates"
         do shell script "cp /Applications/plexconnect_BACKUP/trailers.pem /Applications/PlexConnect/assets/certificates"
@@ -469,6 +476,7 @@ script AppDelegate
                 do shell script "echo not present"
             end if
         end tell
+        do shell script "restartbash.bash"
         tell myProgressBar to stopAnimation:me -- another way
         set animated to false
     end buttonhandlerrestoreall_
@@ -502,8 +510,7 @@ script AppDelegate
         set animated to true
         tell application "Finder"
             if (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
-                quit
-                return
+                do shell script "trashbasebash.bash"
                 else if not (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
                 try
                     do shell script "mkdir /Applications/plexconnect_BACKUP"
@@ -540,8 +547,7 @@ script AppDelegate
         set animated to true
         tell application "Finder"
             if (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
-                quit
-                return
+                do shell script "trashbasebash.bash"
                 else if not (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
                 try
                     do shell script "mkdir /Applications/plexconnect_BACKUP"
@@ -578,8 +584,7 @@ script AppDelegate
         set animated to true
         tell application "Finder"
             if (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
-                quit
-                return
+                do shell script "trashbasebash.bash"
                 else if not (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
                 try
                     do shell script "mkdir /Applications/plexconnect_BACKUP"
@@ -616,8 +621,7 @@ script AppDelegate
         set animated to true
         tell application "Finder"
             if (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
-                quit
-                return
+                do shell script "trashbasebash.bash"
                 else if not (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
                 try
                     do shell script "mkdir /Applications/plexconnect_BACKUP"
@@ -654,8 +658,7 @@ script AppDelegate
         set animated to true
         tell application "Finder"
             if (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
-                quit
-                return
+                do shell script "trashbasebash.bash"
                 else if not (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
                 try
                     do shell script "mkdir /Applications/plexconnect_BACKUP"
@@ -692,8 +695,7 @@ script AppDelegate
         set animated to true
         tell application "Finder"
             if (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
-                quit
-                return
+                do shell script "trashbasebash.bash"
                 else if not (exists folder "Applications:PlexConnect:update:OSX" of the startup disk) then
                 try
                     do shell script "mkdir /Applications/plexconnect_BACKUP"
